Moving getNewCondition out of service into model
